rename camera image query.
